Crate chip8_base

chip8_base provides everything you need to get started building your own CHIP-8 interpreter. See the documentation for the Interpreter trait to get started.

Traits

Interpreter

CHIP-8 interpreters can be built using this trait. step should be implemented on a type representing a CHIP-8 Interpreter to run the interpreter one clock cycle at a time, such that calling it in a loop runs the interpreter.

Functions

run

Starts the interpreter, blocking the current thread and running until killed. Windowing, graphics, sound, and timing are all handled within this method.

Type Definitions

Display

The Interpreter’s representation of the CHIP-8 display. The display is 64x32 pixels, each pixel being either on or off, represented by a 0 or 1, respectively.

Keys

This type is how keyboard input is presented to the Interpreter. Each of the 16 keys can either be down (true) or up (false).